requirements.txt
boto3==1.34.88
fastapi==0.111.0
matplotlib==3.8.3
openai==1.28.0
pandas==2.2.2
python-dotenv==1.0.1
python_docx==1.1.0
Requests==2.31.0
streamlit==1.31.1
uvicorn==0.29.0
